Don't use the "poison sqlite header" logic. could be a problem.
Google gears team implemented a patch to sqlite called "poison the header".
It got included in sqlite android shipped.

This patch is meant to corrupt the database header if some sqlite process
concludes that the database is corrupt. Thereafter, any other sqlite process
opening the same database file would quickly see the corruption and exit.

I am not convinced that this is a useful patch for Android.
And, it patch didn't make it to the open source version of sqlite.

could it be causing previously nonexisting
database corruption, by incorrectly poisoning the database header? thats
what I am suspecting.