when sqlite encounters SQLITE_MISUE error, cause segfault. STOPSHIP.
this CL wil have to be reverted before shipping. thats why the STOPSHIP
comment all over this CL.
goal is to get stacktrace to help debug http://b/2419869 because it is
intermittently occuring bug and sqlite3 doesn't give enough info
when this error occurs.
Richard Hipp of sqlite is going to send a patch that will provide additional
info when this error occurs.