Add PHONEBOOK_COLLATOR toward SQLite so that database users are able to sort based on phonebook in ja_JP locale.
In the future, we may able to use PHONEBOOK_COLLATOR in the other locales, but we don't use
the other phonebook collator now since we cannot estimate how impactful the change is.
Now we have a customized phonebook collator in Japanese and we need it in ContactsProvider,
we'll use only the collator.

Do not use this collator in the other packages and locales now. We may decide revert it in the
near future after furter investigatons.

Need to submit I4dd1b047 to make this change actually effective. Without the change,
we will have a phonebook collator for "previous" locale, not the current one.

Bug: 2514026
Bug: 2373557
Bug: 2373553
Change-Id: If6f548c0a80fe01e779b4dfc46f74224003f1798