get latest sqlite from froyo branch created by richard hipp @ sqlite
this will remove a lot of android specific code from sqlite in
external/sqlite/dist.
we are getting closer to using the public domain sqlite.
 ******* NOTE TO REVIEWERS *********
to compare, don't use gerrit. that will compare this sqlite3.c with
the currently-checkin version, showing you the differences
between this sqlite version and the previous sqlite version.
you may not care about that.
if you just need to see what android changes to sqlite are,
then compare the following two files in this CL
    sqlite3.c.orig
    sqlite3.c