Update patch set 2
Patch Set 2: I would prefer that you didn't submit this

I'm a bit worried about the story behind this change. I'm not quite convinced that we want consumer devices to ship with -ffast-math, and from there I'm not sure that we should have any change that enables that at all.

Patch-set: 2