Update patch set 2
Patch Set 2: No score

If you enable -ffast-math, you better know what you are doing.  If you wants to get comparable results on different platforms, you should not use fast-math.

For 3D graphics, a little loss in accuracy of cross-platform repeatability may not be that big a deal, so using -ffast-math should be okay.  If you want to do high precision physic simulation on your Nexus One, -ffast-math is not your friend.  I read on the internet that SETI@Home is not compiled with -ffast-math because the optimization causes incorrect results.

It is a trade-off.  Do we have any benchmark data to see how much -ffast-math buys us?

BTW, if sqlite3IsNaN is that only thing that is broken by -ffast-math, we can use __builtin_isnan() in lieu of the test in sqlite.

Patch-set: 2