Update patch set 2
Patch Set 2: Do not submit

I don't know much about this flag, but it sounds like it's only okay to turn it on under very controlled circumstances.  So let's not add code that enforces that we don't use it in some specific cases.  If there is a cause to turn it on, it should only be turned on for the targets where it is safe.

Was this change precipitated by something in particular or was it just being careful?

Patch-set: 2