Update patch set 2
Patch Set 2:

I'd definitely feel more comfortable, at least philosophically, if we ended up with per-module opt-in, for areas where we believe that the inaccuracies can be acceptable, especially for well-targeted performance-critical areas. That could be similar to the way some source files are compile as ARM instead of Thumb. That being said, I have no idea which modules would be good candidates, though OpenGL would come to mind. I'm guessing that this should be driven by benchmarks, so that we don't spend time studying the cases that don't being any actual performance improvements.

Patch-set: 2