Support allowing URI authority
From the documentation:
URI filenames normally throw an
error is the authority section is
not either empty or "localhost".
However, if SQLite is compiled with
the SQLITE_ALLOW_URI_AUTHORITY
compile-time option, then the URI is
converted into a Uniform Naming
Convention (UNC) filename and passed
down to the underlying operating system
that way.

Some future versions of SQLite may
change to enable this feature by default.