Cast struct stat st_ino to ino_t before comparing with stored inode
In android, the st_ino field in struct stat is declared as
unsigned long long on 32 bit ABIs, while ino_t is a 32 bit
type. This means that the st_ino field potentially can
express inode values that can't be stored in an ino_t.

As long as Android has run on 32 bit kernels, this hasn't
been any issue, but now when on 64 bit kernels, the actual
inodes can be outside of the 32 bit range (which seems to
happen at least for files within app sandboxes on the /sdcard
partition).

This leads to the fileHasMoved check triggering falsely, since
buf.st_ino isn't equal to the (truncated) stored
pFile->pInode->fileId.ino (which is an ino_t).

By casting buf.st_ino to ino_t before doing the comparison, we
avoid triggering fileHasMoved when the files actually haven't
moved, at the expense of a small theoretical risk of missing
files that actually have moved.

Change-Id: I7e752524d02a033ad62f09b4d8e397d6bec37365