Store inodes in unsigned long long
In 32 bit ABIs, ino_t is a 32 bit type, while the st_ino field
in struct stat is unsigned long long. This means that struct
stat can expose inode numbers that are truncated when stored
in an ino_t.

In practice this has never happened while running on 32 bit
kernels where all inode numbers are 32 bit, but now when
Android runs on 64 bit kernels as well, this seems to happen
for files in app sandboxes on /sdcard.

The fileHasMoved function in sqlite compares the stored
inode value with a new one from stat, and when the stored
value is truncated, this check will falsely indicate that
the file has been moved. When the fileHasMoved function
triggers, other functions start returning errors indicating
that the database is in read-only mode.

(On normal linux, the st_ino field in struct stat is an ino_t.
Therefore, if the kernel ever uses >32 bit inode numbers there,
they are truncated already when exposed via the 32 bit stat
struct, and this issue in sqlite never arises. It only is an
issue on Android where the ABI is slightly different.)

Change-Id: Id79a866c8891d6df47e4eec17ad9eb48e4cb0138