Update patch set 1
Patch Set 1:

Ok, good. Yes, I prefer the other solution as well. Btw, I reported a bug with this as well, at https://code.google.com/p/android/issues/detail?id=79994 (most of the story is the same in the commit message and there, but in case you want to close it once it's fixed).

Patch-set: 1