Update patch set 1
Patch Set 1:

My only general concern about this release
is the increased memory consumption.

I wouldn't think it would be a problem
but if the impact is bigger than expected,
we could modify the makefile to use the old
value for "default lookaside" in the case when
the SVELTE config is used.

Patch-set: 1
Reviewer: Alex Naidis <1092708@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0