Update patch set 1
Patch Set 1:

Is there any way we can create a reproducer which we can send to the sqlite team? I am sure that they would be happy to look into it.

I did a quick check on the open sqlite tickets and this issue was not reported yet.

Patch-set: 1
Reviewer: Alex Naidis <1092708@85c56323-6fa9-3386-8a01-6480fb634889>