Tag FTS3 cursor pointers in fts3ColumnMethod()
FTS3 tables have a magic column, whose name matches the table name, which
is used to obtain a pointer to the cursor for the current operation.
fts3ColumnMethod() now tags its result when used with the magic column,
and fts3FunctionArg() checks the tag and returns an error when asked to
access the magic column on behalf of the user.

Bug: 63852675
Change-Id: Ie0ab625c708d8fcad80db66a373f2684672642b1
Merged-In: Ifb8b0fd23e6306b576c2c947dcab45d811e4becb