Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Owner +1, 

I can't see "Owner +1".

> but I don't understand any implication around it. 

Marking "double_loaded:true" for a shared module doesn't change the build output or its behavior. Instead, it reflects the current state of the library. "Double loading" may happen when a library has a vendor variant and also is a (transitive) dependency of LL-NDK. In this case libsqlite is a VNDK, which means it has a vendor variant and also a system-core variant. And LL-NDK libmediandk also depends on libsqlite. If a certain vendor process depends on both libmediandk and libsqlite, then two variants(vendor and system) of libsqlite may be loaded in the same vendor process, which we call it "double loading".

"Double loadable can be a problem if the library is written with an assumption that it is the only entity in the process that can hold a process-global resource such as mutex, file descriptor, etc." (Quote from jiyong@'s comment https://android-review.googlesource.com/c/platform/external/protobuf/+/898933/1#message-4448b18be3ebd280a4e14d6e4d2d1798c20bb9a0)

Patch-set: 1
Reviewer: Gerrit User 1415877 <1415877@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0