sqlite: Upgrade to SQLite 3.28.0
Downloaded from
https://sqlite.org/2019/sqlite-autoconf-3280000.tar.gz

$ sha1sum sqlite-autoconf-3280000.tar.gz
01b9d8fc77085e144dddc87456c9783e53d09a53  sqlite-autoconf-3280000.tar.gz

dist/orig contains the stock sqlite3 code, as packaged in the tar.gz file above.

dist contains a copy of dist/orig, but with the Android.patch file applied.

The Android specific patch didn't apply cleanly due to merge conflict in shell.c. The conflict was resolved and the patch was regenerated.

More details about the release are available here:
https://www.sqlite.org/releaselog/3_28_0.html

Test: code compiles and device boots with no problems.
Test: select sqlite_version() - returns 3.28.0
Test: atest cts/SQLiteDatabaseTest all passed.
Test: atest cts/SQLSecurityTest
Bug: 135048770
Change-Id: Idf38a3adace9f4bd4feab4012c9cc4de3a1c501c