Update patch set 1
Patch Set 1:

(2 comments)

This commit is part of an experiment and proposed as an alternative to switching sqlite on host back to using libicuuc / libicui18n (see http://r.android.com/957539).

If this is acceptable then it's one less user of external/icu to think about, which would simplify vichang@'s job.

ccross@ / jiyong@ : Any thoughts / suggestions on how to move forward with this? We don't know who might be using the sqlite3 binary on host. The libsqlite is a bit more obvious but also it's unclear whether they need any of the behavior that ICU "unlocks".

Patch-set: 1
Reviewer: Gerrit User 1038039 <1038039@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0