Use SQLITE_ENABLE_ICU build flag for host Linux and Mac
This makes the host variant of the 'sqlite' target consistent with the
Android variant. When SQLITE_ENABLE_ICU is defined, several SQLite ICU
extensions are enabled for host Linux and Mac, such as the `regexp`
function.

See https://sqlite.org/src/dir/ext/icu for more details.

Host SQLite is used in tools like Robolectric and LayoutLib.

All host platforms (Linux, Mac, and Windows) support ICU. There will be
a follow-up CL to enable SQLITE_ENABLE_ICU for Windows, but the ICU
dependency on Windows needs to be updated.

Also enable a host variant for 'libsqlite_static_noicu' so binaries such
as py2-launcher and perfetto will not need to include ICU libraries.

Test: Robolectric sqlite database tests
Bug: 308173442
Change-Id: I6a8de8f5719856dff173ca3e79c2e8afad748243