Reorganize external/sqlite for androidx
This CL reorganizes the folder structure of `external/sqlite` that is independent of platform branches and meant to be used for `androidx.sqlite:sqlite-bundled`. It removes platform specifics and only keeps sqlite3 sources from the amalgamation and enables androidx to upgrade the version of SQLite used at its own pace.

Test: ./frameworks/support/sqlite/scripts/runConformanceTest.sh
Change-Id: Ide6aa1ba3e3ed5e9fdb969fa3be86388c2dd0020