Allow multiple releases of sqlite
This change allows the build system to hold multiple versions of the
sqlite library.  This change only allows the multiple versions to be
stored concurrently.  It does not change the build process.

A sqlite version is stored in the directory dist/<year>/<tarball-name>
under external/sqlite.  This naming corresponds to how tarballs are
saved on sqlite.org.  Note that the year of the release is important.

To create a new sqlite version for 3.42.0 (created in 2023):
  cd external/sqlite
  ./UPDATE-SOURCE.bash 2023 3.42.0

The update process downloads the sqlite source and applies the
Android.patch file. If the patch is not clean, fix the sqlite files by
hand, regenerate the patch file, and rerun the update-source command:
  ./REBUILD-ANDROID-PATCH.bash 2023 3.44.2
  ./UPDATE-SOURCE.bash 2023 3.42.0

The patch should apply cleanly.

Test: manual inspection

Bug: 323176655

Change-Id: Ia61ba6890dec1ea1a8c450b70d85b964d679d21a