Consume build flag in sqlite Android.bp
The two Android.bp files for sqlite now use the build flag
RELEASE_PACKAGE_LIBSQLITE3.  The value of the build flag must be a
seven-digit sqlite release number.  For example, release 3.44.2 is
coded as 3440200.  This is the same number that sqlite.org uses for
the release URLs.

The default source files (also 3.42.0) have been copied into the
directory external/sqlite/dist/sqlite-default.  These files are used
for every target that is not aware of trunk-stable build flags.

Bug: 323176655

Change-Id: Iba1dd75408f42b35eb52b680c11a9947aad04f3f