Update the sqlite build scripts
UPDATE-SOURCE now accepts a url on the command line.  This is
important when the canonical url does not work (as is the case when
using an unreleased sqlite tarball).

The version number can now be specified as four integers, which is
useful when using an unreleased sqlite tarball.

Some common functions used by UPDATE-SOURCE and REBUILD-PATCH have
been moved into shared file.

Test: The updated scripts were used to install an unreleased sqlite
tarball from the 3.44.2 train.

Bug: 323176655
Change-Id: I674f9079407df1e9e00541cbed92041f0ded2f3e