Minor tweaks to the sqlite upgrade script
This changes the sqlite upgrade script so that the the year is
provided through a switch and is only needed if the full URL to the
patch is not specified.  (The year is needed to form a standard
sqlite.org URL from the version.)  The year now defaults to the
current year.

Manually tested with a full URL and no year, and with a version+year.

Flag: EXEMPT bug-fix
Bug: 397436943
Test: manual

Change-Id: I422a0df36d5baf237ff0b6e942cb8a4f8d352515