Install sqlite 3.50.1 source files
This installs sqlite 3.50.1 source files.  It does not include the new
version of sqlite in the build.  The actual version of sqlite that is
installed is controlled by a trunk-stable build flag, which is not
modified in this CL.

This also deletes the 3.50.0 source files.  These source files failed
to build under one of the Android targets.

Flag: EXEMPT source install only
Bug: 422156354
Test: presubmit
Change-Id: I53e6d7c7a3fe5172c8b41313c6d85ce717ddbd43