Install sqlite 3.44.5 source files
This installs sqlite 3.44.5 source files.  It does not include the new
version of sqlite in the build.  The actual version of sqlite that is
installed is controlled by a trunk-stable build flag, which is not
modified in this CL.

Note: the installation on main is primarily to create a CL that can be
cherry-picked to older releases.  The current plan of record is that
main will adopt 3.50.x.

Flag: EXEMPT source install only
Bug: 430889718
Bug: 281064726
Test: presubmit
Change-Id: I6e7cf1fcfe2c1e6a432cfb73fb45c3bee0cb36d3