Automate publishing a new sqlite release
This script is used to accomplish two things when a new sqlite release
moves to "next".  The METADATA file for sqlite is updated and the
sqlite-default symlink is updated.

We must remember to run this script and commit the resulting CLs when
RELEASE_PACKAGE_LIBSQLITE3 changes its value in "next".

The instructions for upgrading SQLite are updated to refer to this new
script.  The instructions also now discuss backporting SQLite updates
to earlier Android releases when, for example, a security patch must
be applied.

Flag: EXEMPT bug-fix
Bug: 430910716
Test: manually ran the script
Change-Id: I337159a2d7ef39fce78a2d2bc379289d339ad26a