Allow sqlite updates from downloaded tarballs
The sqlite script UPDATE-SOURCE.bash takes a URL as a parameter, which
the script uses to download a tarball from sqlite.org.  Recent
security changes in sqlite.org make this difficult.  This changes to
script to accept a file along with the URL, so that we can
download the tarball manually and still use the automation scripts.

Flag: EXEMPT helper script
Bug: 447138765
Test: manual - the script was used to install a new version of sqlite
from a tarball that had been manually downloaded.

Change-Id: I2cfaddc6da657a99e13fc84a7f37430913007ea6